requirements.txt
Flask==3.0.0
supabase-py==2.4.0
requests==2.32.3
ytmusicapi==1.9.0
gunicorn==21.2.0
